feat(): Move Layout component to views dir
