+++ resources/js/views/Menus/Index.vue
@@ -54,7 +54,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
     export default {
         props: {
             menus: Array
