+++ resources/js/views/Pages/Create.vue
@@ -65,7 +65,7 @@
 
 <script>
     import _ from 'lodash';
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
     import MarkdownField from '../../components/fields/MarkdownField';
 
     export default {
