+++ resources/js/views/Pages/Edit.vue
@@ -64,7 +64,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
     import MarkdownField from '../../components/fields/MarkdownField';
 
     export default {
