+++ resources/js/views/Pages/Index.vue
@@ -135,7 +135,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
 
     export default {
         props: {
