+++ resources/js/views/Posts/Edit.vue
@@ -72,7 +72,7 @@
 
 <script>
     import _ from 'lodash';
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
     import MarkdownField from "../../components/fields/MarkdownField";
     import TagsField from "../../components/fields/TagsField";
