+++ resources/js/views/Posts/Index.vue
@@ -1,5 +1,6 @@
 <template>
     <layout title="Posts">
+        <BaseButton></BaseButton>
         <div class="bg-white px-4 py-5 border-b border-gray-200 sm:px-6">
             <div class="-ml-4 -mt-2 flex items-center justify-between flex-wrap sm:flex-no-wrap">
                 <div class="ml-4 mt-2">
@@ -68,7 +69,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
     export default {
         props: {
             posts: Array
