+++ resources/js/views/Profile/Show.vue
@@ -54,7 +54,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
 
     export default {
         components: {
