+++ resources/js/views/Users/Create.vue
@@ -78,7 +78,7 @@
 
 <script>
     import _ from 'lodash';
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
 
     export default {
         components: {
