+++ resources/js/views/Users/Edit.vue
@@ -77,7 +77,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
 
     export default {
         components: {
