+++ resources/js/views/Users/Show.vue
@@ -52,7 +52,7 @@
 </template>
 
 <script>
-    import Layout from '../../Layout';
+    import Layout from '../Layout';
 
     export default {
         components: {
